Fix several year-2038-problems
See https://en.wikipedia.org/wiki/Year_2038_problem

With this change, the ciff format supports timestamps between 1970 and 2106.

This patch was done while reviewing potential year-2038 issues in openSUSE.

+++ src/x3f/x3f_parse_process.cpp
@@ -160,7 +160,7 @@ void LibRaw::parse_x3f()
         if (!strcmp(name, "WB_DESC"))
           strcpy(imgdata.color.model2, value);
         if (!strcmp(name, "TIME"))
-          imgdata.other.timestamp = atoi(value);
+          imgdata.other.timestamp = atoll(value);
         if (!strcmp(name, "SHUTTER"))
           imgdata.other.shutter = float(atof(value));
         if (!strcmp(name, "APERTURE"))
